exp double

+++ src/AZTERKETA2/Azterketa.java
@@ -67,10 +67,10 @@ public class Azterketa {
             Evaluation eval = new Evaluation(data);
             weka.classifiers.functions.SMO smo = new SMO();
             double fmax =0;
-            int expmax=0;
+            double expmax=0;
             weka.classifiers.functions.supportVector.PolyKernel pk = new PolyKernel();
             System.out.println("Exponente maximoa aldatu da, orain da: "+expmax);
-            for(int i = 1; i<6; i++){
+            for(double i = 1; i<6; i++){
                 pk.setExponent(i);
                 smo.setKernel(pk);
